+++ NEWS.md
@@ -5,6 +5,7 @@
 * Permutational test added for trajectoryDirectionality().
 * New function trajectoryMetrics() for evaluation of multiple metrics.
 * New function trajectoryWindowMetrics() for metrics evaluated over moving windows.
+* New function trajectoryShifts() to compare trajectories that are similar in shape but differ in speed.
 
 # ecotraj 0.2.1
 * Update of function 'centerTrajectories' to add flexibility in trajectory centroid definition.